Update django.po
Hey Brian,

I've done all the changes and (hopefully) done all the new translations. After today, I will be on holiday for two weeks so can check the app after that, or if it needs to be done beforehand then ask Laura to contact me by phone and I can do it sooner!